deletation of machine instance object
