+++ packages/jam/jam-host-calls/refine/expunge.ts
@@ -2,8 +2,7 @@ import { type HostCallHandler, type PvmExecution, tryAsHostCallIndex } from "@ty
 import { type GasCounter, type Registers, tryAsSmallGas } from "@typeberry/pvm-interpreter";
 import { LegacyHostCallResult } from "../results";
 import { CURRENT_SERVICE_ID } from "../utils";
-import { tryAsMachineId } from "./machine-instance";
-import type { RefineExternalities } from "./refine-externalities";
+import { type RefineExternalities, tryAsMachineId } from "./refine-externalities";
 
 const IN_OUT_REG = 7;
