+++ packages/jam/jam-host-calls/refine/peek.ts
@@ -10,8 +10,7 @@ import {
 import { assertNever } from "@typeberry/utils";
 import { LegacyHostCallResult } from "../results";
 import { CURRENT_SERVICE_ID } from "../utils";
-import { tryAsMachineId } from "./machine-instance";
-import { PeekPokeError, type RefineExternalities } from "./refine-externalities";
+import { PeekPokeError, type RefineExternalities, tryAsMachineId } from "./refine-externalities";
 
 const IN_OUT_REG = 7;
