+++ packages/jam/jam-host-calls/refine/zero.test.ts
@@ -6,8 +6,7 @@ import { MemoryBuilder, Registers, gasCounter, tryAsGas } from "@typeberry/pvm-i
 import { tryAsSbrkIndex } from "@typeberry/pvm-interpreter/memory/memory-index";
 import { OK, Result } from "@typeberry/utils";
 import { LegacyHostCallResult } from "../results";
-import { type MachineId, tryAsMachineId } from "./machine-instance";
-import { NoMachineError } from "./refine-externalities";
+import { type MachineId, NoMachineError, tryAsMachineId } from "./refine-externalities";
 import { TestRefineExt } from "./refine-externalities.test";
 import { Zero } from "./zero";
